requirements.txt
streamlit
google - generative api
python-dotenv
langchain
pyPDF2
chromadb